Moved SimpleLogger to jda pkg, made it package private, removed jdocs
