+++ src/main/java/net/dv8tion/jda/core/utils/SimpleLogger.java
@@ -23,12 +23,7 @@
  *
  */
 
-/**
- * Changes made to original Loggerger from SLF4J
- * - Made constructor public
- */
-
-package org.slf4j.impl;
+package net.dv8tion.jda.core.utils;
 
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
@@ -41,87 +36,13 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Properties;
 
-import org.slf4j.Logger;
 import org.slf4j.helpers.FormattingTuple;
 import org.slf4j.helpers.MarkerIgnoringBase;
 import org.slf4j.helpers.MessageFormatter;
 import org.slf4j.helpers.Util;
 import org.slf4j.spi.LocationAwareLogger;
 
-/**
- * <p>Simple implementation of {@link Logger} that sends all enabled log messages,
- * for all defined loggers, to the console ({@code System.err}).
- * The following system properties are supported to configure the behavior of this logger:</p>
- *
- * <ul>
- * <li><code>org.slf4j.Loggerger.logFile</code> - The output target which can be the <em>path</em> to a file, or
- * the special values "System.out" and "System.err". Default is "System.err".
- *
- * <li><code>org.slf4j.Loggerger.defaultLogLevel</code> - Default log level for all instances of Loggerger.
- * Must be one of ("trace", "debug", "info", "warn", or "error"). If not specified, defaults to "info". </li>
- *
- * <li><code>org.slf4j.Loggerger.log.<em>a.b.c</em></code> - Logging detail level for a Loggerger instance
- * named "a.b.c". Right-side value must be one of "trace", "debug", "info", "warn", or "error". When a Loggerger
- * named "a.b.c" is initialized, its level is assigned from this property. If unspecified, the level of nearest parent
- * logger will be used, and if none is set, then the value specified by
- * <code>org.slf4j.Loggerger.defaultLogLevel</code> will be used.</li>
- *
- * <li><code>org.slf4j.Loggerger.showDateTime</code> - Set to <code>true</code> if you want the current date and
- * time to be included in output messages. Default is <code>false</code></li>
- *
- * <li><code>org.slf4j.Loggerger.dateTimeFormat</code> - The date and time format to be used in the output messages.
- * The pattern describing the date and time format is defined by
- * <a href="http://docs.oracle.com/javase/1.5.0/docs/api/java/text/SimpleDateFormat.html"><code>SimpleDateFormat</code></a>.
- * If the format is not specified or is invalid, the number of milliseconds since start up will be output. </li>
- *
- * <li><code>org.slf4j.Loggerger.showThreadName</code> -Set to <code>true</code> if you want to output the current
- * thread name. Defaults to <code>true</code>.</li>
- *
- * <li><code>org.slf4j.Loggerger.showLogName</code> - Set to <code>true</code> if you want the Logger instance name
- * to be included in output messages. Defaults to <code>true</code>.</li>
- *
- * <li><code>org.slf4j.Loggerger.showShortLogName</code> - Set to <code>true</code> if you want the last component
- * of the name to be included in output messages. Defaults to <code>false</code>.</li>
- *
- * <li><code>org.slf4j.Loggerger.levelInBrackets</code> - Should the level string be output in brackets? Defaults
- * to <code>false</code>.</li>
- *
- * <li><code>org.slf4j.Loggerger.warnLevelString</code> - The string value output for the warn level. Defaults
- * to <code>WARN</code>.</li>
-
- * </ul>
- *
- * <p>In addition to looking for system properties with the names specified above, this implementation also checks for
- * a class loader resource named <code>"Loggerger.properties"</code>, and includes any matching definitions
- * from this resource (if it exists).</p>
- *
- * <p>With no configuration, the default output includes the relative time in milliseconds, thread name, the level,
- * logger name, and the message followed by the line separator for the host.  In log4j terms it amounts to the "%r [%t]
- * %level %logger - %m%n" pattern. </p>
- * <p>Sample output follows.</p>
- * <pre>
- * 176 [main] INFO examples.Sort - Populating an array of 2 elements in reverse order.
- * 225 [main] INFO examples.SortAlgo - Entered the sort method.
- * 304 [main] INFO examples.SortAlgo - Dump of integer array:
- * 317 [main] INFO examples.SortAlgo - Element [0] = 0
- * 331 [main] INFO examples.SortAlgo - Element [1] = 1
- * 343 [main] INFO examples.Sort - The next log statement should be an error message.
- * 346 [main] ERROR examples.SortAlgo - Tried to dump an uninitialized array.
- *   at org.log4j.examples.SortAlgo.dump(SortAlgo.java:58)
- *   at org.log4j.examples.Sort.main(Sort.java:64)
- * 467 [main] INFO  examples.Sort - Exiting main method.
- * </pre>
- *
- * <p>This implementation is heavily inspired by
- * <a href="http://commons.apache.org/logging/">Apache Commons Logging</a>'s Logger.</p>
- *
- * @author Ceki G&uuml;lc&uuml;
- * @author <a href="mailto:sanders@apache.org">Scott Sanders</a>
- * @author Rod Waldhoff
- * @author Robert Burrell Donkin
- * @author C&eacute;drik LIME
- */
-public class SimpleLogger extends MarkerIgnoringBase {
+class SimpleLogger extends MarkerIgnoringBase {
 
     private static final long serialVersionUID = -632788891211436180L;
     private static final String CONFIGURATION_FILE = "Loggerger.properties";
@@ -149,7 +70,6 @@ public class SimpleLogger extends MarkerIgnoringBase {
     private static boolean LEVEL_IN_BRACKETS = false;
     private static String WARN_LEVEL_STRING = "WARN";
 
-    /** All system properties used by <code>Loggerger</code> start with this prefix */
     public static final String SYSTEM_PREFIX = "org.slf4j.Loggerger.";
 
     public static final String DEFAULT_LOG_LEVEL_KEY = SYSTEM_PREFIX + "defaultLogLevel";
@@ -184,9 +104,6 @@ public class SimpleLogger extends MarkerIgnoringBase {
         return (prop == null) ? defaultValue : "true".equalsIgnoreCase(prop);
     }
 
-    // Initialize class attributes.
-    // Load properties file, if found.
-    // Override with system properties.
     static void init() {
         INITIALIZED = true;
         loadProperties();
@@ -254,15 +171,10 @@ public class SimpleLogger extends MarkerIgnoringBase {
         }
     }
 
-    /** The current log level */
     protected int currentLogLevel = LOG_LEVEL_INFO;
-    /** The short name of this simple log instance */
     private transient String shortLogName = null;
 
-    /**
-     * Public Loggerger constructor.
-     */
-    public SimpleLogger(String name) {
+    SimpleLogger(String name) {
         if (!INITIALIZED) {
             init();
         }
@@ -304,14 +216,6 @@ public class SimpleLogger extends MarkerIgnoringBase {
         return LOG_LEVEL_INFO;
     }
 
-    /**
-     * This is our internal implementation for logging regular (non-parameterized)
-     * log messages.
-     *
-     * @param level   One of the LOG_LEVEL_XXX constants defining the log level
-     * @param message The message itself
-     * @param t       The exception whose stack trace should be logged
-     */
     private void log(int level, String message, Throwable t) {
         if (!isLevelEnabled(level)) {
             return;
@@ -399,14 +303,6 @@ public class SimpleLogger extends MarkerIgnoringBase {
         return name.substring(name.lastIndexOf(".") + 1);
     }
 
-    /**
-     * For formatted messages, first substitute arguments and then log.
-     *
-     * @param level
-     * @param format
-     * @param arg1
-     * @param arg2
-     */
     private void formatAndLog(int level, String format, Object arg1, Object arg2) {
         if (!isLevelEnabled(level)) {
             return;
@@ -415,13 +311,6 @@ public class SimpleLogger extends MarkerIgnoringBase {
         log(level, tp.getMessage(), tp.getThrowable());
     }
 
-    /**
-     * For formatted messages, first substitute arguments and then log.
-     *
-     * @param level
-     * @param format
-     * @param arguments a list of 3 ore more arguments
-     */
     private void formatAndLog(int level, String format, Object... arguments) {
         if (!isLevelEnabled(level)) {
             return;
@@ -430,223 +319,128 @@ public class SimpleLogger extends MarkerIgnoringBase {
         log(level, tp.getMessage(), tp.getThrowable());
     }
 
-    /**
-     * Is the given log level currently enabled?
-     *
-     * @param logLevel is this level enabled?
-     */
     protected boolean isLevelEnabled(int logLevel) {
         // log level are numerically ordered so can use simple numeric
         // comparison
         return (logLevel >= currentLogLevel);
     }
 
-    /** Are {@code trace} messages currently enabled? */
     public boolean isTraceEnabled() {
         return isLevelEnabled(LOG_LEVEL_TRACE);
     }
 
-    /**
-     * A simple implementation which logs messages of level TRACE according
-     * to the format outlined above.
-     */
     public void trace(String msg) {
         log(LOG_LEVEL_TRACE, msg, null);
     }
 
-    /**
-     * Perform single parameter substitution before logging the message of level
-     * TRACE according to the format outlined above.
-     */
     public void trace(String format, Object param1) {
         formatAndLog(LOG_LEVEL_TRACE, format, param1, null);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * TRACE according to the format outlined above.
-     */
     public void trace(String format, Object param1, Object param2) {
         formatAndLog(LOG_LEVEL_TRACE, format, param1, param2);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * TRACE according to the format outlined above.
-     */
     public void trace(String format, Object... argArray) {
         formatAndLog(LOG_LEVEL_TRACE, format, argArray);
     }
 
-    /** Log a message of level TRACE, including an exception. */
     public void trace(String msg, Throwable t) {
         log(LOG_LEVEL_TRACE, msg, t);
     }
 
-    /** Are {@code debug} messages currently enabled? */
     public boolean isDebugEnabled() {
         return isLevelEnabled(LOG_LEVEL_DEBUG);
     }
 
-    /**
-     * A simple implementation which logs messages of level DEBUG according
-     * to the format outlined above.
-     */
     public void debug(String msg) {
         log(LOG_LEVEL_DEBUG, msg, null);
     }
 
-    /**
-     * Perform single parameter substitution before logging the message of level
-     * DEBUG according to the format outlined above.
-     */
     public void debug(String format, Object param1) {
         formatAndLog(LOG_LEVEL_DEBUG, format, param1, null);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * DEBUG according to the format outlined above.
-     */
     public void debug(String format, Object param1, Object param2) {
         formatAndLog(LOG_LEVEL_DEBUG, format, param1, param2);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * DEBUG according to the format outlined above.
-     */
     public void debug(String format, Object... argArray) {
         formatAndLog(LOG_LEVEL_DEBUG, format, argArray);
     }
 
-    /** Log a message of level DEBUG, including an exception. */
     public void debug(String msg, Throwable t) {
         log(LOG_LEVEL_DEBUG, msg, t);
     }
 
-    /** Are {@code info} messages currently enabled? */
     public boolean isInfoEnabled() {
         return isLevelEnabled(LOG_LEVEL_INFO);
     }
 
-    /**
-     * A simple implementation which logs messages of level INFO according
-     * to the format outlined above.
-     */
     public void info(String msg) {
         log(LOG_LEVEL_INFO, msg, null);
     }
 
-    /**
-     * Perform single parameter substitution before logging the message of level
-     * INFO according to the format outlined above.
-     */
     public void info(String format, Object arg) {
         formatAndLog(LOG_LEVEL_INFO, format, arg, null);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * INFO according to the format outlined above.
-     */
     public void info(String format, Object arg1, Object arg2) {
         formatAndLog(LOG_LEVEL_INFO, format, arg1, arg2);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * INFO according to the format outlined above.
-     */
     public void info(String format, Object... argArray) {
         formatAndLog(LOG_LEVEL_INFO, format, argArray);
     }
 
-    /** Log a message of level INFO, including an exception. */
     public void info(String msg, Throwable t) {
         log(LOG_LEVEL_INFO, msg, t);
     }
 
-    /** Are {@code warn} messages currently enabled? */
     public boolean isWarnEnabled() {
         return isLevelEnabled(LOG_LEVEL_WARN);
     }
 
-    /**
-     * A simple implementation which always logs messages of level WARN according
-     * to the format outlined above.
-     */
     public void warn(String msg) {
         log(LOG_LEVEL_WARN, msg, null);
     }
 
-    /**
-     * Perform single parameter substitution before logging the message of level
-     * WARN according to the format outlined above.
-     */
     public void warn(String format, Object arg) {
         formatAndLog(LOG_LEVEL_WARN, format, arg, null);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * WARN according to the format outlined above.
-     */
     public void warn(String format, Object arg1, Object arg2) {
         formatAndLog(LOG_LEVEL_WARN, format, arg1, arg2);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * WARN according to the format outlined above.
-     */
     public void warn(String format, Object... argArray) {
         formatAndLog(LOG_LEVEL_WARN, format, argArray);
     }
 
-    /** Log a message of level WARN, including an exception. */
     public void warn(String msg, Throwable t) {
         log(LOG_LEVEL_WARN, msg, t);
     }
 
-    /** Are {@code error} messages currently enabled? */
     public boolean isErrorEnabled() {
         return isLevelEnabled(LOG_LEVEL_ERROR);
     }
 
-    /**
-     * A simple implementation which always logs messages of level ERROR according
-     * to the format outlined above.
-     */
     public void error(String msg) {
         log(LOG_LEVEL_ERROR, msg, null);
     }
 
-    /**
-     * Perform single parameter substitution before logging the message of level
-     * ERROR according to the format outlined above.
-     */
     public void error(String format, Object arg) {
         formatAndLog(LOG_LEVEL_ERROR, format, arg, null);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * ERROR according to the format outlined above.
-     */
     public void error(String format, Object arg1, Object arg2) {
         formatAndLog(LOG_LEVEL_ERROR, format, arg1, arg2);
     }
 
-    /**
-     * Perform double parameter substitution before logging the message of level
-     * ERROR according to the format outlined above.
-     */
     public void error(String format, Object... argArray) {
         formatAndLog(LOG_LEVEL_ERROR, format, argArray);
     }
 
-    /** Log a message of level ERROR, including an exception. */
     public void error(String msg, Throwable t) {
         log(LOG_LEVEL_ERROR, msg, t);
     }
